requirements.txt
cryptography==45.0.3
fastapi==0.115.12
Pillow==11.2.1
pydantic==2.11.5
PyJWT==2.10.1
Requests==2.32.3
